# doc/api.yaml
openapi: 3.0.0
info: 
    title: WASAPhoto
    description: WASAPhoto app backend REST API
    version: 1.1.0
...
paths: 
...
  /photos/{photoId}:
    parameters: 
      - $ref: "#/components/parameters/photoId"
    delete: 
      tags: ["photos"]
      operationId: deletePhoto
      summary: deletes a post from profile
      description: |
          deletes a photo with given photoId from your list of photos. can not be applied to other profiles
      responses: 
        "204": 
          description: photo deleted sucessfully
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404": 
          description: photo not found 
        "500": {$ref: "#/components/responses/internalServerError"}
    get: 
      tags: ["photos"]
      operationId: getPhoto
      summary: get a photo with given photoId
      description: |
            returns a photo with given photoId
      responses: 
        "200":
          description: returns photo sucessfully
          content: 
            application/png:
              schema:  
                {$ref: "#/components/schemas/photoBody"}
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404": 
            description: user not found 
        "500": {$ref: "#/components/responses/internalServerError"}
  /followings/{userId}:
    parameters:
       - $ref: "#/components/parameters/userId"
    get: 
      tags: ["followings"]
      operationId: getFollowings
      summary: get list of followings of user with given userId
      description: |
            returns a list of followings of a user with given userId
      responses: 
        "200":
          description: returns followings sucessfully
          content: 
            application/json: 
              schema: 
                {$ref: "#/components/schemas/users"}
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404": 
            description: user not found 
        "500": {$ref: "#/components/responses/internalServerError"}
    put:
      tags: ["followings"]
      operationId: followUser
      summary: follow user with given username
      description: |
          add user with given userId to your followings list. can not be applied to other users followers 
      responses:
        "204":
          description: followed user sucessfully
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404":
          description: user not found
        "500": {$ref: "#/components/responses/internalServerError"}
    delete:
      tags: ["followings"] 
      operationId: unfollowUser
      summary: deletes user from your following list
      description: |
          delete a user with given userId from your following list. can not be applied to other user's following list
      responses:
        "204": 
          description: unfollowed user sucessfully 
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404": 
          description: user not found
        "500": {$ref: "#/components/responses/internalServerError"}
  /followers/{userId}:
    parameters:
       - $ref: "#/components/parameters/userId"
    get: 
      tags: ["followers"]
      operationId: getFollowers
      summary: get list of followers of user with given userId
      description: |
            returns a list of followers of a user with given userId
      responses: 
        "200":
          description: returns followers sucessfully
          content: 
            application/json: 
              schema: 
                {$ref: "#/components/schemas/users"}
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404": 
            description: user not found 
        "500": {$ref: "#/components/responses/internalServerError"}
  /banned/{userId}:
   parameters:
      - $ref: "#/components/parameters/userId"
   get: 
        tags: ["bans"]
        operationId: getBanned
        summary: get list of profiles of all users that you banned
        description: |
             returns list of users that you added to your banned list, can only be used to see your own bans list. 
        responses: 
          "200":
            description: returns banned users sucessfully
            content: 
              application/json: 
                schema: 
                 {$ref: "#/components/schemas/banned"}
          "400": {$ref: "#/components/responses/badRequest"}
          "401": {$ref: "#/components/responses/unauthorized"}
          "404": 
              description: bans list not found 
          "500": {$ref: "#/components/responses/internalServerError"}
   put:
      tags: ["bans"]
      operationId: banUser
      summary: ban user with given userId
      description: |
          adds user with given userId to your banned list. can not be applied to other users followers
      responses:
        "204":
          description: banned user sucessfully
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404":
          description: user not found
        "500": {$ref: "#/components/responses/internalServerError"}
   delete: 
      tags: ["bans"]
      operationId: unbanUser
      summary: remove user from bans list
      description: |
           remove user from your bans list. can not be applied to other users bans list
      responses: 
        "204": 
          description: user unbanned sucessfully
        "400": {$ref: "#/components/responses/badRequest"}
        "401": {$ref: "#/components/responses/unauthorized"}
        "404": 
          description: user not found
        "500": {$ref: "#/components/responses/internalServerError"}
...
components:
    parameters: 
...
       userId:
          name: userId
          in: path
          description: |
            unique identification number of a user. can not be changed
          schema: 
            type: integer
            readOnly: true    
          required: true      
       photoId:
        description: unique identification number of a photo
        schema: 
          type: integer
        name: photoId
        in: path
        required: true
    schemas:
        id: 
          title: id
          description: |
                unique id of an object
          type: integer
          maximum: 1234
          minimum: 1
          readOnly: true
        username: 
          title: username
          description: |
              unique username of a user. unlike the userId it            can be changed 
          type: string
          pattern: ".*"
          example: maria
          maxLength: 20
          minLength: 3
        users:
          title: collection of user profiles 
          description: | 
                array of users that will be returned in repsonse 
          type: array
          properties:
            number_of_users: 
              description: number of users in list
              readOnly: true
              type: integer
              minimum: 0
          minItems: 1234
          maxItems: 1234
          items:
              $ref: "#/components/schemas/username"
...
        banned: 
          title: list of banned users
          description:  list of profiles that you added to your bans list
          type: array
          properties:
            number_of_bans: 
              description: number of users in bans list
              type: integer
              minimum: 0
          minItems: 0
          maxItems: 1234
          items:
              $ref: "#/components/schemas/id"
...
        photoBody: 
          title: data representing the photo 
          description: |
              data of the image that is stored separetly from 
              other photo data such as caption, date and ect.
          type: string
          format: binary
          maxLength: 167777160
          minLength: 124
...
    responses: 
      badRequest: 
        description: |
          malformed request syntax, invalid request message framing, or deceptive request routing
      internalServerError: 
        description: |
          server encountered an unexpected condition that prevented it from fulfilling the request
      unauthorized: 
        description: |
          client request has not been completed because it lacks valid authentication credentials for the requested resource
    securitySchemes:
      userIdAuth: 
        type: http
        scheme: bearer
security:
  - userIdAuth: []